requirements-dev.txt
pytest
Sphinx
sphinx_rtd_theme
numpydoc
travis-sphinx
flake8
